requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
redis==5.1.1
kafka-python==2.0.2
spacy==3.7.6
transformers==4.44.2
sentence-transformers==3.1.1
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
cryptography==43.0.1
slowapi==0.1.9
pyjwt==2.9.0
httpx==0.27.2
feedparser==6.0.11
aiohttp==3.10.8
faiss-cpu==1.8.0.post1
geopy==2.4.1
numpy==1.26.4
prometheus-client==0.21.0
structlog==24.4.0
langdetect==1.0.9
pytest==8.3.3
pytest-asyncio==0.24.0

# OpenClaw integration (uses existing httpx)
# No extra deps; bridge uses httpx from main requirements

# Training (optional)
datasets>=2.14.0
scikit-learn>=1.3.0
accelerate>=0.24.0
psutil>=5.9.0
